refine gatedgcn
